requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
python-dotenv==1.0.1
Werkzeug==3.0.1
Jinja2==3.1.3
SQLAlchemy==2.0.25
WTForms==3.1.1
email-validator==2.1.0.post1
Pillow==10.2.0
pytest==8.0.0
pytest-cov==4.1.0
black==24.1.1
flake8==7.0.0
isort==5.13.2
mypy==1.8.0 